fix: click event 로직 수정
- carNamesEvent, racingCountEvent 클릭 이벤트 로직 index.js로 수정
